:memo: Docs: Add jsdocs for state and predeploy packages
