+++ vm/predeploys/docs/functions/definePredeploy.md
@@ -8,6 +8,8 @@
 
 > **definePredeploy**\<`TName`, `THumanReadableAbi`\>(`__namedParameters`): [`Predeploy`](../classes/Predeploy.md)\<`TName`, `THumanReadableAbi`\>
 
+Defines a predeploy contract to use in the tevm vm
+
 ## Type parameters
 
 ▪ **TName** extends `string`
@@ -18,9 +20,33 @@
 
 ▪ **\_\_namedParameters**: `Pick`\<[`Predeploy`](../classes/Predeploy.md)\<`TName`, `THumanReadableAbi`\>, `"contract"` \| `"address"`\>
 
+## Returns
+
+## Example
+
+```ts
+import { definePredeploy } from 'tevm/predeploys'
+import { createMemoryTevm } from 'tevm/vm'
+import { createScript } from 'tevm/contract'
+
+const predeploy = definePredeploy({
+  address: `0x${'23'.repeat(20)}`,
+  contract: createScript({
+    name: 'PredeployExample',
+    humanReadableAbi: ['function foo() external pure returns (uint256)'],
+    bytecode: '0x608060405234801561001057600080fd5b5061012f806100206000396000f3fe608060405260043610610041576000357c0100',
+    deployedBytecode: '0x608060405260043610610041576000357c010000
+  })
+})
+
+const vm = createMemoryTevm({
+ predeploys: [predeploy.predeploy()],
+})
+```
+
 ## Source
 
-[definePredeploy.ts:26](https://github.com/evmts/tevm-monorepo/blob/main/vm/predeploys/src/definePredeploy.ts#L26)
+[definePredeploy.ts:27](https://github.com/evmts/tevm-monorepo/blob/main/vm/predeploys/src/definePredeploy.ts#L27)
 
 ***
 Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
