+++ vm/predeploys/src/index.ts
@@ -1,1 +1,3 @@
-export * from './definePredeploy.js'
+export { definePredeploy } from './definePredeploy.js'
+export type { Predeploy } from './Predeploy.js'
+export type { CustomPredeploy } from './CustomPredeploy.js'
