+++ vm/state/docs/interfaces/TevmStateManagerInterface.md
@@ -18,7 +18,7 @@
 
 #### Source
 
-[vm/state/src/TevmStateManager.ts:34](https://github.com/evmts/tevm-monorepo/blob/main/vm/state/src/TevmStateManager.ts#L34)
+[vm/state/src/TevmStateManager.ts:36](https://github.com/evmts/tevm-monorepo/blob/main/vm/state/src/TevmStateManager.ts#L36)
 
 ***
