+++ vm/state/src/index.ts
@@ -1,3 +1,7 @@
-export * from './TevmStateManager.js'
-export * from './SerializableState.js'
-export * from './DefaultTevmStateManager.js'
+export {
+	TevmStateManager,
+	type TevmStateManagerOpts,
+	type TevmStateManagerInterface,
+} from './TevmStateManager.js'
+export type { SerializableTevmState } from './SerializableState.js'
+export { DefaultTevmStateManager } from './DefaultTevmStateManager.js'
